+++ src/components/page-blog/page-blog.tsx
@@ -97,8 +97,6 @@ export class PageBlog {
     let topicString = item.tracks[0];
     for (var i = 1; i < item.tracks.length; i++) {
       topicString = topicString + ", " + item.tracks[i];
-
-
     }
     return topicString;
   }
@@ -138,7 +136,6 @@ export class PageBlog {
                     {item.title}
                     <p innerHTML={item.teaser}></p>
                     <p><em>Posted {new Date(item.datePublished).toDateString()}, Tagged {this.createTopicList(item)}
-
                     </em></p>
                   </ion-label>
                 </ion-item>