+++ changelog.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Added** [#67](https://github.com/codyburleson/cburleson-pwa/issues/67) Page: eclipse-tip-format-source-code-on-save
 - **Added** [#61](https://github.com/codyburleson/cburleson-pwa/issues/61) Page: better-error-messages-from-gulp-using-gulp-util
 - **Added** [#62](https://github.com/codyburleson/cburleson-pwa/issues/62) Page: binding-to-a-zsphere-armature
+- **Added** Page: user-impersonation-programming-in-websphere-portal
+- **Added** [#63](https://github.com/codyburleson/cburleson-pwa/issues/63) Page: book-review-more-than-everything-by-vanessa-foster
 
 ## [1.1.3] - 2019-12-26